server/requirements.txt
fastapi
uvicorn[standard]
pyjwt
python-dotenv
